alteracoes gerais; rodar alguns experimentos

+++ src/C/bruno/GeneticOpenCL/genetica.h
@@ -2,9 +2,9 @@
 #define GENETICA_H_INCLUDED
 #include "arvore.h"
 
-void inicializaCheia(Arvore pop[], int num, int indice, int* conjuntoOpTerm, int NUM_OPBIN, int NUM_OPUN, int N, int* seed, int maxDados, int minDados);
-void inicializaAleatorio(Arvore pop[], int num, int indice, int* conjuntoOpTerm, int NUM_OPBIN, int NUM_OPUN, int N, int* seed, int maxDados, int minDados);
-void inicializaPopulacao(Arvore pop[], int* conjuntoOpTerm, int NUM_OPBIN, int NUM_OPUN, int N, int* seed, int maxDados, int minDados);
+void inicializaCheia(Arvore pop[], int num, int indice, int* conjuntoOpTerm, int NUM_OPBIN, int NUM_OPUN, int N, int* seed, float maxDados, float minDados);
+void inicializaAleatorio(Arvore pop[], int num, int indice, int* conjuntoOpTerm, int NUM_OPBIN, int NUM_OPUN, int N, int* seed, float maxDados, float minDados);
+void inicializaPopulacao(Arvore pop[], int* conjuntoOpTerm, int NUM_OPBIN, int NUM_OPUN, int N, int* seed, float maxDados, float minDados);
